service/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-dotenv>=1.0

sqlalchemy>=2.0
alembic>=1.13

python-jose[cryptography]>=3.3
python-multipart>=0.0.9

numpy<2
pillow>=10.0

diffusers==0.33.0
transformers>=4.41
accelerate>=0.31
safetensors>=0.4
huggingface-hub>=0.27,<1.0
tokenizers>=0.19
sentencepiece>=0.2.0
